Tests/LibWeb/Text/expected/HTML/referrer-policy-http-header.txt

no-referrer policy: PASS (no Referer header sent)
origin policy: PASS (only origin sent)
